karigar-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langgraph==0.0.40
langchain==0.1.0
langchain-openai==0.0.5
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
chromadb==0.4.18
tiktoken==0.5.2
python-dotenv==1.0.0
geopy==2.4.0

# Image Processing
Pillow==10.1.0

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Payments
razorpay==1.4.1

# Notifications
twilio==8.10.0
firebase-admin==6.3.0

# Other tools
reportlab
qrcode
